Fix merge files problems
